Added some changes
high_volume_filter now is more restrictive.
Added new momentum strategy based simulation

+++ check_high_volume_breakers.pl
@@ -1,4 +1,5 @@
 #! /usr/bin/perl
+
 use lib qw( ..);
 use DBI;
 use DateTime qw(:all);
@@ -6,7 +7,14 @@ use DateTime qw(:all);
 use strict;
 use warnings;
 
+my $simulation = 0;
+my $send_email = 1;
+my $send_beep = 0;
+
 sub check_for_download_time {
+	if ($simulation) {
+		return 1;
+	}
 	my $time_now = DateTime->now( time_zone => 'local' );
 	my $download = 1;
 	my $dow = $time_now->day_of_week;
@@ -23,7 +31,7 @@ sub check_for_download_time {
 			$download = 0;
 		} else {
 			if ($time_now->hour == 9) {
-				if ($time_now->minute < 25) {
+				if ($time_now->minute < 30) {
 					$download = 0;
 				}
 			} else {
@@ -66,7 +74,9 @@ my $userid = "";
 my $password = "";
 
 my $repeat_always = 1;
-my $delay_in_seconds = 300;
+my $delay_in_seconds = $simulation == 1? 20: 300;
+
+my $email_program = "mutt";
 
 my @data;
 my %prev3_price;
@@ -80,9 +90,24 @@ my %prev_vol;
 my %cur_price;
 my %cur_vol;
 my %per_change_in_price;
+my %time_of_last_price;
+my %date_of_last_price;
+my $lot_size_in_cash = 200000;
+my $max_per_lot_size_to_vol = 5;
+
+
 while ($repeat_always) {
+	my $start_time = time();
+
+	#my $time_now = DateTime->now( time_zone => 'local' );
+	#print "<h6>$time_now</h6>\n";
+
 	if (check_for_download_time()) {
-		system("cp 01realtime_data.db check_volume.db");
+		if ($simulation) {
+			system("cp realtime_copy.db check_volume.db");
+		} else {
+			system("cp 01realtime_data.db check_volume.db");
+		}
 		#print "Copied<br>\n";
 
 		my $dbh = DBI->connect($dsn, $userid, $password, { RaiseError => 1 })
@@ -133,10 +158,13 @@ while ($repeat_always) {
 		 $cur_price{$data[1]} = $data[5];
 		 $cur_vol{$data[1]} = $data[9];
 		 $per_change_in_price{$data[1]} = $data[8];
+		 $time_of_last_price{$data[1]} = $data[12];
+		 $date_of_last_price{$data[1]} = $data[13];
 		}
 
 		$dbh->disconnect();
 		#print "Closed database successfully<br>\n";
+		print "Time: $time_of_last_price{'ADANIENT'}<br>\n";
 
 		my @change_price;
 		my @change_volume;
@@ -169,12 +197,63 @@ while ($repeat_always) {
 			my $per_change_pr1 = $change_price[1] == 0? 0 : ($change_price[2]/$change_price[1])*100.0;
 			my $per_change_pr0 = $change_price[0] == 0? 0 : ($change_price[1]/$change_price[0])*100.0;
 
-			if ($per_change_vol0 > 3000.0 || $per_change_vol1 > 3000.0 || $per_change_vol2 > 3000.0) {
-				printf("<b><a href=\"http://chartink.com/stocks/%s.html\">%s</a></b>(%.02f:%.02f%%): %.02f%% (<b>%.02f%%</b>) %.02f%% (<b>%.02f%%</b>) %.02f%% (<b>%.02f%%</b>)<br>\n",$stock_name,$stock_name,$cur_price{$stock_name},$per_change_in_price{$stock_name},$per_change_pr0,$per_change_vol0,$per_change_pr1,$per_change_vol1,$per_change_pr2,$per_change_vol2);
+			my $vol_per_to_lot_size2 = $change_volume[3] == 0? $max_per_lot_size_to_vol : ((($lot_size_in_cash/$cur_price{$stock_name})/$change_volume[3]) - 1)*100.0;
+			my $vol_per_to_lot_size1 = $change_volume[2] == 0? $max_per_lot_size_to_vol : ((($lot_size_in_cash/$cur_price{$stock_name})/$change_volume[2]) - 1)*100.0;
+			my $vol_per_to_lot_size0 = $change_volume[1] == 0? $max_per_lot_size_to_vol : ((($lot_size_in_cash/$cur_price{$stock_name})/$change_volume[1]) - 1)*100.0;
+
+			if (($vol_per_to_lot_size2 < $max_per_lot_size_to_vol) && ($vol_per_to_lot_size1 < $max_per_lot_size_to_vol) &&
+				($vol_per_to_lot_size0 < $max_per_lot_size_to_vol) &&
+				($per_change_in_price{$stock_name} > 1.0) && ($per_change_pr2 > 0) && ($change_price[3] > 0) && ($change_price[2] > 0) && ($change_price[1] > 0) &&
+				($change_volume[0] > 5000) && ($change_volume[1] > 5000) && ($change_volume[2] > 5000) && ($change_volume[3] > 5000) &&
+				($per_change_vol0 > 1000.0 || $per_change_vol1 > 1000.0 || $per_change_vol2 > 1000.0)) {
+				printf("<b><a href=\"http://chartink.com/stocks/%s.html\">%s</a></b>(%.02f:%.02f%%): %.02f[%.02f%%] (<b>%d[%.02f%%]</b>) %.02f[%.02f%%] (<b>%d[%.02f%%]</b>) %.02f[%.02f%%] (<b>%d[%.02f%%]</b>)<br>\n",$stock_name,$stock_name,$cur_price{$stock_name},$per_change_in_price{$stock_name},$change_price[1],$per_change_pr0,$change_volume[1],$per_change_vol0,$change_price[2],$per_change_pr1,$change_volume[2],$per_change_vol1,$change_price[3],$per_change_pr2,$change_volume[3],$per_change_vol2);
+				my $beep_sent = $send_beep == 1? system("./beep.sh > /dev/null") : 0;
+				my $email_cmd = qq(-s "Fourways Profit: $time_of_last_price{$stock_name}: Buy $stock_name \($cur_price{$stock_name}\)" mimansoor\@gmail.com,lksingh74\@gmail.com < /dev/null);
+				my $email_sent = $send_email == 1? system("$email_program $email_cmd") : 0;
+
+				#Now store it in a DB with Buy, take profit and stop loss values
+				my $driver   = "SQLite";
+				my $database = "high_volume_calls.db";
+				my $dsn = "DBI:$driver:dbname=$database";
+				my $userid = "";
+				my $password = "";
+				my $dbh = DBI->connect($dsn, $userid, $password, { RaiseError => 1 })
+						      or die $DBI::errstr;
+				#Get some unique id value
+				my $insert_id = time();
+				my $insert_stmt = qq(INSERT INTO high_volume_calls (ID, NAME, DATE, TIME, BUY_PRICE)
+							VALUES ($insert_id, '$stock_name', '$date_of_last_price{$stock_name}', '$time_of_last_price{$stock_name}', $cur_price{$stock_name}));
+				my $rv = $dbh->do($insert_stmt) or warn print "$insert_stmt\n";
+				$dbh->disconnect();
 			}
 		}
 	}
-	sleep $delay_in_seconds;
-	my $time_now = DateTime->now( time_zone => 'local' );
-	print "<h6>$time_now</h6>\n";
+
+to_sleep:
+	my $end_time = time();
+	my $work_time = ($end_time - $start_time);
+
+	#To ensure alignment with actual clock if drift
+	#in seconds is > delay/2 (seconds)
+	my $rounding = $end_time % $delay_in_seconds;
+	#print "Rounding: $rounding\n";
+	if ($rounding > $delay_in_seconds/2) {
+		$rounding = $delay_in_seconds - $rounding;
+	} else {
+		#Decrease time from sleep
+		$rounding *= -1.0;
+	}
+
+	#Adjust the work time to keep getting every delay(300) seconds
+	#And if required move to the next round minute.
+	my $sleep_time = $delay_in_seconds - $work_time + $rounding;
+
+	#If for some reason sleep goes negative,
+	#Sleep at least delay seconds
+	if ($sleep_time <= 0) {
+		$sleep_time = $delay_in_seconds;
+	}
+
+	#print "Sleeping for $sleep_time\n";
+	sleep $sleep_time;
 }
